move ArticleContextMenuPresenting into separate file
